Detect when no nested in a VM.
When not in a VM, `this.parent()` in not a function, so check
for that, and return a default claass (primary) if that's the
case. It's also possible that the parent VM renders a layout
which requires the parent to choose the class, in which case
the parent's `buttonLayout` method would return a class. If
this method exists, let it override.