Ability to render a full-width CTA.
It's common for CTAs to render as a fullwidth callout. If `isFullWidth`
is true, it will render as a full-width CTA. This can be passed in as
an option on the data context. Also added an onRendered callback
which will compare the width of the CTA's container to the width of the
body, and automatically render as fullWidth if they are about the same
width.